chore: move poll helpers to utils
